requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
docutils==0.16
flit==3.2.0
flit-core==3.2.0
idna==2.10
mypy==0.812
mypy-extensions==0.4.3
nose2==0.10.0
pathspec==0.8.1
Pygments==2.8.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
rich==10.0.0
six==1.15.0
sockets==1.0.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
